chore(ci): manager integration tests
