+++ tests/integration/kind/openebs/static/static.go
@@ -1,0 +1,8 @@
+package static
+
+import (
+	"embed"
+)
+
+//go:embed *
+var FS embed.FS
