--- tests/integration/openebs/static/static.go
@@ -1,8 +1,0 @@
-package static
-
-import (
-	"embed"
-)
-
-//go:embed *
-var FS embed.FS
